set variant and some refactor
